✨ change to hybrid_group commands
