+++ app/api/routers/users.py
@@ -1,8 +1,8 @@
 from fastapi import APIRouter, HTTPException
 from pydantic import BaseModel
 
+from app.core.mediator import Mediator
 from app.core.result import is_err
-from app.mediator import Mediator
 from app.usecases.users.create_user import CreateUserCommand
 from app.usecases.users.get_user import GetUserQuery
