+++ app/usecases/memberships/request_join_team.py
@@ -5,13 +5,13 @@ from dataclasses import dataclass
 
 from injector import inject
 
+from app.core.mediator import Request, RequestHandler
 from app.core.result import Err, Ok, Result, is_err
 from app.domain.aggregates.team import Team
 from app.domain.aggregates.team_membership import TeamMembership
 from app.domain.aggregates.user import User
 from app.domain.repositories import IUnitOfWork
 from app.domain.value_objects import TeamId, UserId
-from app.mediator import Request, RequestHandler
 from app.usecases.result import ErrorType, UseCaseError
 
 logger = logging.getLogger(__name__)
